Changed model format for day_of_week #54
If the day of week is stored as a three letter English abbreviation, a
problem could occur when python's locale is changed. The strftime('%a')
could suddenly return Russian abbreviations or who knows. So now the day
of week is stored as an integer (strftime('%w')). Now at least the
Russians can use Freevle too =D